fix cas characterId invalide

+++ controller/charactersController.js
@@ -22,6 +22,9 @@ exports.add = async function (req, res) {
 exports.update = async function (req, res) {
   const { characterId } = req.params;
   const character = await Character.findByPk(characterId);
+  if (!character) {
+    return res.status(404).send("characterId invalide");
+  }
   Character.update(req.body, { where: { id: character.id } })
     .then((e) => res.status(200).send(e))
     .catch((e) => res.status(500).send(e));
@@ -30,11 +33,14 @@ exports.update = async function (req, res) {
 exports.delete = async function (req, res) {
   const { characterId } = req.params;
   const character = await Character.findByPk(characterId);
+  if (!character) {
+    return res.status(404).send("characterId invalide");
+  }
   Character.destroy({
     where: {
       id: character.id,
     },
   })
-    .then((e) => res.status(200).send(e))
+    .then(() => res.sendStatus(200))
     .catch((e) => res.status(500).send(e));
 };
